refactor test cases for scheduling heuristics
